fix(provider): harden tool request handling
